triangular distribution

+++ fms_pwflat.h
@@ -26,7 +26,7 @@ namespace fms::pwflat {
 	// f(u) = NaN, u < 0
 	// f(u) = _f, u > t[n-1]
 	template<class T = double, class F = double>
-	inline F value(T u, unsigned n, const T* t, const F* f,
+	inline F value(T u, size_t n, const T* t, const F* f,
 		F _f = NaN<F>)
 	{
 		if (u < 0) {
